requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.77.1
anyio==4.12.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
distro==1.9.0
Django==5.2.11
django-celery-beat==2.8.1
django-colorfield==0.14.0
django-cors-headers==4.9.0
django-filter==25.2
django-tenants==3.9.0
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
docstring_parser==0.17.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
kombu==5.6.2
lxml==6.0.2
Markdown==3.10.1
packaging==26.0
pillow==12.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
vk_api==11.10.0
wcwidth==0.5.3
